Move back button inside sidebar to fix weird animation
