+++ osu.Game/Overlays/SettingsSubPanel.cs
@@ -1,17 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-#nullable disable
-
 using osu.Framework.Allocation;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Sprites;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
-using osu.Game.Overlays.Settings;
-using osuTK;
 
 namespace osu.Game.Overlays
 {
@@ -25,69 +15,8 @@ namespace osu.Game.Overlays
         [BackgroundDependencyLoader]
         private void load()
         {
-            AddInternal(new BackButton
-            {
-                Anchor = Anchor.BottomLeft,
-                Origin = Anchor.BottomLeft,
-                Action = Hide
-            });
         }
 
         protected override bool DimMainContent => false; // dimming is handled by main overlay
-
-        public partial class BackButton : SidebarButton
-        {
-            private Drawable content;
-
-            public BackButton()
-                : base(HoverSampleSet.Default)
-            {
-            }
-
-            [BackgroundDependencyLoader]
-            private void load()
-            {
-                Size = new Vector2(SettingsSidebar.EXPANDED_WIDTH);
-
-                Padding = new MarginPadding(40);
-
-                AddRange(new Drawable[]
-                {
-                    content = new FillFlowContainer
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Direction = FillDirection.Vertical,
-                        AutoSizeAxes = Axes.Both,
-                        Spacing = new Vector2(5),
-                        Children = new Drawable[]
-                        {
-                            new SpriteIcon
-                            {
-                                Anchor = Anchor.TopCentre,
-                                Origin = Anchor.TopCentre,
-                                Size = new Vector2(30),
-                                Shadow = true,
-                                Icon = FontAwesome.Solid.ChevronLeft
-                            },
-                            new OsuSpriteText
-                            {
-                                Anchor = Anchor.TopCentre,
-                                Origin = Anchor.TopCentre,
-                                Font = OsuFont.GetFont(size: 16, weight: FontWeight.Regular),
-                                Text = @"back",
-                            },
-                        }
-                    }
-                });
-            }
-
-            protected override void UpdateState()
-            {
-                base.UpdateState();
-
-                content.FadeColour(IsHovered ? ColourProvider.Light1 : ColourProvider.Light3, FADE_DURATION, Easing.OutQuint);
-            }
-        }
     }
 }
